List of applications now populated using a json configuration file. Also updated the urls.py in anticipation of the deprecation in the new django versions
