+++ integral_view/core/replication/urls.py
@@ -3,8 +3,7 @@ from django.contrib.auth.decorators import login_required
 
 from integral_view.core.replication.views.remote_replication_management import create_remote_replication, view_remote_replications, delete_remote_replication, update_remote_replication, update_rsync_remote_replication_pause_schedule, update_remote_replication_user_comment
 
-
-urlpatterns = patterns('',
+urlpatterns = [
                        url(r'^$', login_required(view_remote_replications)),
 
                        # From views/remote_replication_management.py
@@ -20,5 +19,5 @@ urlpatterns = patterns('',
                            login_required(update_rsync_remote_replication_pause_schedule)),
                        url(r'^delete_remote_replication/',
                            login_required(delete_remote_replication)),
-                        )
+]
 # vim: tabstop=8 softtabstop=0 expandtab ai shiftwidth=4 smarttab
