+++ integral_view/core/system/templates/update_manifest.html
@@ -12,7 +12,7 @@
     <input type="text" value="yes" class="form-control" name="action" style="display:none">
     <div class="btn-group btn-group-sm" role="group" aria-label="...">
       <button class="btn btn-primary cover-page">Yes</button>
-      <a class="btn btn-default" href="/view_system_info/">No</a>
+      <a class="btn btn-default" href="/system/view_system_info/">No</a>
     </div>
   </form>
   <br>
