+++ integral_view/settings.py
@@ -12,8 +12,27 @@ https://docs.djangoproject.com/en/1.8/ref/settings/
 
 from integralstor import config
 platform_root, err = config.get_platform_root()
-STATIC_DIR_PATH = '%s/integral_view/static' % platform_root
-TEMPLATE_DIR_PATH_LIST = ["%s/integral_view/templates"%platform_root, "%s/integral_view/core/storage/templates"%platform_root, "%s/integral_view/core/replication/templates"%platform_root, "%s/integral_view/core/tasks/templates"%platform_root, "%s/integral_view/core/keys_certs/templates"%platform_root, "%s/integral_view/core/networking/templates"%platform_root, "%s/integral_view/core/users_groups/templates"%platform_root, "%s/integral_view/core/storage_access/templates"%platform_root, "%s/integral_view/core/system/templates"%platform_root, "%s/integral_view/core/monitoring/templates"%platform_root, "%s/integral_view/core/application_management/templates"%platform_root, "%s/integral_view/applications/urbackup/templates"%platform_root, "%s/integral_view/applications/storage_insights/templates"%platform_root]
+
+TEMPLATE_DIR_PATH_LIST = ["%s/integral_view/templates"%platform_root, "%s/integral_view/core/storage/templates"%platform_root, "%s/integral_view/core/replication/templates"%platform_root, "%s/integral_view/core/tasks/templates"%platform_root, "%s/integral_view/core/keys_certs/templates"%platform_root, "%s/integral_view/core/networking/templates"%platform_root, "%s/integral_view/core/users_groups/templates"%platform_root, "%s/integral_view/core/storage_access/templates"%platform_root, "%s/integral_view/core/system/templates"%platform_root, "%s/integral_view/core/monitoring/templates"%platform_root, "%s/integral_view/core/application_management/templates"%platform_root]
+
+import os
+ROOT_PATH = os.path.dirname(__file__)
+STATIC_DIR_PATH_LIST = [os.path.join(ROOT_PATH, 'static')]
+
+applications, err = config.get_applications_config()
+application_template_dirs = []
+application_static_dirs = []
+if applications:
+    for base_dir, app in applications.items():
+        application_template_dirs.append('%s/integral_view/applications/%s/templates'%(platform_root, base_dir))
+        application_static_dirs.append('%s/integral_view/static/%s'%(platform_root, base_dir))
+
+if application_template_dirs:
+    TEMPLATE_DIR_PATH_LIST.extend(application_template_dirs)
+if application_static_dirs:
+    STATIC_DIR_PATH_LIST.extend(application_static_dirs)
+
+
 LOGIN_URL = '/login/'
 
 # Build paths inside the project like this: os.path.join(BASE_DIR, ...)
@@ -111,11 +130,6 @@ USE_TZ = True
 STATIC_URL = '/static/'
 
 # Additional locations of static files
-STATICFILES_DIRS = (
-    # Put strings here, like "/home/html/static" or "C:/www/django/static".
-    # Always use forward slashes, even on Windows.
-    # Don't forget to use absolute paths, not relative paths.
-    STATIC_DIR_PATH,
-)
+STATICFILES_DIRS = STATIC_DIR_PATH_LIST
 
 # vim: tabstop=8 softtabstop=0 expandtab ai shiftwidth=4 smarttab
