+++ integral_view/urls.py
@@ -7,8 +7,17 @@ from django.contrib import admin
 admin.autodiscover()
 from integral_view.views.admin_auth import login, logout, update_admin_password
 
-urlpatterns = patterns('',
+from integralstor import config
+applications, err = config.get_applications_config()
+app_patterns = []
+if applications:
+    for app_tag in applications.keys():
+        try:
+            app_patterns.append(url(r'^applications/%s/'%app_tag, include('integral_view.applications.%s.urls'%app_tag)))
+        except Exception, e:
+            pass
 
+urlpatterns = [
                        # Uncomment the admin/doc line below to enable admin documentation:
                        # url(r'^admin/doc/', include('django.contrib.admindocs.urls')),
 
@@ -24,8 +33,6 @@ urlpatterns = patterns('',
                        url(r'^users_groups/', include('integral_view.core.users_groups.urls')),
                        url(r'^keys_certs/', include('integral_view.core.keys_certs.urls')),
                        url(r'^storage_access/', include('integral_view.core.storage_access.urls')),
-                       url(r'^applications/urbackup/', include('integral_view.applications.urbackup.urls')),
-                       url(r'^applications/storage_insights/', include('integral_view.applications.storage_insights.urls')),
                        url(r'^system/', include('integral_view.core.system.urls')),
 
                        # From views/admin_auth.py
@@ -33,10 +40,11 @@ urlpatterns = patterns('',
                        url(r'^logout/', logout, name="logout"),
                        url(r'^$', login),
                        url(r'^update_admin_password/',
-                           login_required(update_admin_password)),
+                           login_required(update_admin_password))
+]
 
-
-                       )
+if app_patterns:
+    urlpatterns.extend(app_patterns)
 
 
 # vim: tabstop=8 softtabstop=0 expandtab ai shiftwidth=4 smarttab
